Fixed race condition where "setInput" was causing a race condition.
Having setInput within pre_detect meant that it could have been called
multiple times prior to a single call to forward, meaning frames would
be dropped. This was causing throttle_node to not decrease, resulting in
a deadlock of the system. By moving the call to "setInput" from
"pre_detect" into "detect", the race condition is removed.

I don't think there's really any other sensible way of dealing with
this, but for some strange reason, when Matt tried replacing std::bind
with lambdas, the race condition no longer held true. I attempted the
same, but couldn't replicate Matt's outcome. Perhaps he was doing
something with the lambdas that essentially synchronised "pre_detect"
and "detect" nodes.

+++ src/tracker.cpp
@@ -120,7 +120,7 @@ void Tracker::merge(const Detections &detections)
         float best_merginess = 0;
         Track* merge_track;
 
-        for (int i = 0; i < tracks.size(); i++)
+        for (size_t i = 0; i < tracks.size(); i++)
         {
             Track* track = tracks[i];
             float iou = IoU(track->box, detection.box);
